refactor: create transfer action component
