+++ app/src/components/sections/ExpenseAccountView/MyApprovedExpenseSection.vue
@@ -1,6 +1,6 @@
 <template>
   <CardComponent title="My Approved Expense">
-    <div v-if="expenseAccountEip712Address && newExpenseData">
+    <div v-if="newExpenseData">
       <!-- TODO display this only if the use have an approved expense -->
       <!-- Expense A/c Info Section -->
       <!-- New Header -->
@@ -9,7 +9,7 @@
         :columns="columns"
       >
         <template #action-data="{ row }">
-          <ButtonUI
+          <!--<ButtonUI
             variant="success"
             :disabled="!expenseDataStore.myApprovedExpenses || row.status !== 'enabled'"
             v-if="true"
@@ -21,7 +21,8 @@
             "
             data-test="transfer-button"
             >Spend</ButtonUI
-          >
+          >-->
+          <TransferAction :row="row as ExpenseResponse" />
         </template>
         <template #expiryDate-data="{ row }">
           <span>{{ new Date(Number(row.data.expiry) * 1000).toLocaleString('en-US') }}</span>
@@ -42,7 +43,7 @@
         </template>
       </TableComponent>
 
-      <ModalComponent v-model="transferModal">
+      <!-- <ModalComponent v-model="transferModal">
         <TransferForm
           v-if="
             transferModal && getTokens(newExpenseData, signatureToTransfer, balances).length > 0
@@ -58,7 +59,7 @@
           "
           @closeModal="() => (transferModal = false)"
         />
-      </ModalComponent>
+      </ModalComponent> -->
     </div>
   </CardComponent>
 </template>
@@ -69,21 +70,22 @@ import { computed, ref, watch } from 'vue'
 import type { BudgetLimit, BudgetData, ExpenseResponse } from '@/types'
 import { USDC_ADDRESS } from '@/constant'
 import CardComponent from '@/components/CardComponent.vue'
-import TransferForm from '@/components/forms/TransferForm.vue'
-import ModalComponent from '@/components/ModalComponent.vue'
+// import TransferForm from '@/components/forms/TransferForm.vue'
+// import ModalComponent from '@/components/ModalComponent.vue'
 import { useUserDataStore, useToastStore, useTeamStore, useExpenseDataStore } from '@/stores'
 import { parseError, log, tokenSymbol, getCurrentUserExpenses, getTokens } from '@/utils'
 import { useWriteContract, useWaitForTransactionReceipt } from '@wagmi/vue'
 import { estimateGas } from '@wagmi/core'
 import expenseAccountABI from '@/artifacts/abi/expense-account-eip712.json'
 import { type Address, parseEther, zeroAddress, encodeFunctionData, type Abi } from 'viem'
-import ButtonUI from '@/components/ButtonUI.vue'
+// import ButtonUI from '@/components/ButtonUI.vue'
 import ERC20ABI from '@/artifacts/abi/erc20.json'
 import { readContract } from '@wagmi/core'
 import { config } from '@/wagmi.config'
 import TableComponent, { type TableColumn } from '@/components/TableComponent.vue'
 import { useContractBalance, useTanstackQuery } from '@/composables'
 import type { TokenId } from '@/constant'
+import TransferAction from './TransferAction.vue'
 //#endregion
 
 const columns = [
@@ -119,19 +121,19 @@ const transferModal = ref(false)
 const tokenAmount = ref('')
 const tokenRecipient = ref('')
 const signatureToTransfer = ref('')
-const transferData = ref({
-  address: { name: '', address: '' },
-  token: { symbol: '', balance: 0, tokenId: '' as TokenId },
-  amount: '0'
-})
+// const transferData = ref({
+//   address: { name: '', address: '' },
+//   token: { symbol: '', balance: 0, tokenId: '' as TokenId },
+//   amount: '0'
+// })
 //#endregion
 
 const teamStore = useTeamStore()
 const currentUserAddress = useUserDataStore().address
 const expenseDataStore = useExpenseDataStore()
-const { balances } = useContractBalance(
-  ref(teamStore.getContractAddressByType('ExpenseAccountEIP712'))
-)
+// const { balances } = useContractBalance(
+//   ref(teamStore.getContractAddressByType('ExpenseAccountEIP712'))
+// )
 
 const {
   data: newExpenseData
@@ -147,207 +149,201 @@ const {
   }
 )
 
-//#region Computed Values
-const expenseAccountEip712Address = computed(
-  () =>
-    teamStore.currentTeam?.teamContracts.find(
-      (contract) => contract.type === 'ExpenseAccountEIP712'
-    )?.address as Address
-)
-//#endregion
+// //#region Computed Values
+// const expenseAccountEip712Address = computed(
+//   () =>
+//     teamStore.currentTeam?.teamContracts.find(
+//       (contract) => contract.type === 'ExpenseAccountEIP712'
+//     )?.address as Address
+// )
+// //#endregion
 
-//#region Composables
-const { addErrorToast, addSuccessToast } = useToastStore()
+// //#region Composables
+// const { addErrorToast, addSuccessToast } = useToastStore()
 
-//expense account transfer
-const {
-  writeContract: executeExpenseAccountTransfer,
-  isPending: isLoadingTransfer,
-  error: errorTransfer,
-  data: transferHash
-} = useWriteContract()
+// //expense account transfer
+// const {
+//   writeContract: executeExpenseAccountTransfer,
+//   isPending: isLoadingTransfer,
+//   error: errorTransfer,
+//   data: transferHash
+// } = useWriteContract()
 
-const { isLoading: isConfirmingTransfer, isSuccess: isConfirmedTransfer } =
-  useWaitForTransactionReceipt({
-    hash: transferHash
-  })
+// const { isLoading: isConfirmingTransfer, isSuccess: isConfirmedTransfer } =
+//   useWaitForTransactionReceipt({
+//     hash: transferHash
+//   })
 
-// Token approval
-const { writeContract: approve, error: approveError, data: approveHash } = useWriteContract()
+// // Token approval
+// const { writeContract: approve, error: approveError, data: approveHash } = useWriteContract()
 
-const { isLoading: isConfirmingApprove, isSuccess: isConfirmedApprove } =
-  useWaitForTransactionReceipt({
-    hash: approveHash
-  })
-//#endregion
+// const { isLoading: isConfirmingApprove, isSuccess: isConfirmedApprove } =
+//   useWaitForTransactionReceipt({
+//     hash: approveHash
+//   })
+// //#endregion
 
-//#region Functions
-const transferFromExpenseAccount = async (to: string, amount: string) => {
-  tokenAmount.value = amount
-  tokenRecipient.value = to
+// //#region Functions
+// const transferFromExpenseAccount = async (to: string, amount: string) => {
+//   tokenAmount.value = amount
+//   tokenRecipient.value = to
 
-  const budgetLimit = expenseDataStore.myApprovedExpenses.find(
-    (item) => item.signature === signatureToTransfer.value
-  ) as BudgetLimit
+//   const budgetLimit = expenseDataStore.myApprovedExpenses.find(
+//     (item) => item.signature === signatureToTransfer.value
+//   ) as BudgetLimit
 
-  if (expenseAccountEip712Address.value && expenseDataStore.myApprovedExpenses) {
-    if (budgetLimit.tokenAddress === zeroAddress) await transferNativeToken(to, amount, budgetLimit)
-    else await transferErc20Token()
-  }
-}
+//   if (expenseAccountEip712Address.value && expenseDataStore.myApprovedExpenses) {
+//     if (budgetLimit.tokenAddress === zeroAddress) await transferNativeToken(to, amount, budgetLimit)
+//     else await transferErc20Token()
+//   }
+// }
 
-const transferNativeToken = async (to: string, amount: string, budgetLimit: BudgetLimit) => {
-  try {
-    const args = [
-      to,
-      parseEther(amount),
-      {
-        ...budgetLimit,
-        budgetData: budgetLimit.budgetData.map((item) => ({
-          ...item,
-          value:
-            item.budgetType === 0
-              ? item.value
-              : budgetLimit.tokenAddress === zeroAddress
-                ? parseEther(`${item.value}`)
-                : BigInt(Number(item.value) * 1e6)
-        }))
-      },
-      signatureToTransfer.value
-    ]
-    const data = encodeFunctionData({
-      abi: expenseAccountABI,
-      functionName: 'transfer',
-      args
-    })
-    await estimateGas(config, {
-      to: expenseAccountEip712Address.value,
-      data
-    })
-    executeExpenseAccountTransfer({
-      address: expenseAccountEip712Address.value,
-      args,
-      abi: expenseAccountABI,
-      functionName: 'transfer'
-    })
-  } catch (error) {
-    console.error('Error in transferNativeToken:', parseError(error, expenseAccountABI as Abi))
-    log.error('Error in transferNativeToken:', parseError(error, expenseAccountABI as Abi))
-    addErrorToast(parseError(error, expenseAccountABI as Abi))
-    transferERC20loading.value = false
-    isLoadingTransfer.value = false
-  }
-}
-const transferERC20loading = ref(false)
-// Token transfer function
-const transferErc20Token = async () => {
-  if (
-    !expenseAccountEip712Address.value ||
-    !tokenAmount.value ||
-    !tokenRecipient.value ||
-    !expenseDataStore.myApprovedExpenses
-  )
-    return
+// const transferNativeToken = async (to: string, amount: string, budgetLimit: BudgetLimit) => {
+//   try {
+//     const args = [
+//       to,
+//       parseEther(amount),
+//       {
+//         ...budgetLimit,
+//         budgetData: budgetLimit.budgetData.map((item) => ({
+//           ...item,
+//           value:
+//             item.budgetType === 0
+//               ? item.value
+//               : budgetLimit.tokenAddress === zeroAddress
+//                 ? parseEther(`${item.value}`)
+//                 : BigInt(Number(item.value) * 1e6)
+//         }))
+//       },
+//       signatureToTransfer.value
+//     ]
+//     const data = encodeFunctionData({
+//       abi: expenseAccountABI,
+//       functionName: 'transfer',
+//       args
+//     })
+//     await estimateGas(config, {
+//       to: expenseAccountEip712Address.value,
+//       data
+//     })
+//     executeExpenseAccountTransfer({
+//       address: expenseAccountEip712Address.value,
+//       args,
+//       abi: expenseAccountABI,
+//       functionName: 'transfer'
+//     })
+//   } catch (error) {
+//     console.error('Error in transferNativeToken:', parseError(error, expenseAccountABI as Abi))
+//     log.error('Error in transferNativeToken:', parseError(error, expenseAccountABI as Abi))
+//     addErrorToast(parseError(error, expenseAccountABI as Abi))
+//     transferERC20loading.value = false
+//     isLoadingTransfer.value = false
+//   }
+// }
+// const transferERC20loading = ref(false)
+// // Token transfer function
+// const transferErc20Token = async () => {
+//   if (
+//     !expenseAccountEip712Address.value ||
+//     !tokenAmount.value ||
+//     !tokenRecipient.value ||
+//     !expenseDataStore.myApprovedExpenses
+//   )
+//     return
 
-  const tokenAddress = USDC_ADDRESS
+//   const tokenAddress = USDC_ADDRESS
 
-  transferERC20loading.value = true
-  const _amount = BigInt(Number(tokenAmount.value) * 1e6)
+//   transferERC20loading.value = true
+//   const _amount = BigInt(Number(tokenAmount.value) * 1e6)
 
-  const budgetLimit = expenseDataStore.myApprovedExpenses.find(
-    (item) => item.signature === signatureToTransfer.value
-  ) as BudgetLimit
+//   const budgetLimit = expenseDataStore.myApprovedExpenses.find(
+//     (item) => item.signature === signatureToTransfer.value
+//   ) as BudgetLimit
 
-  const allowance = await readContract(config, {
-    address: tokenAddress as Address,
-    abi: ERC20ABI,
-    functionName: 'allowance',
-    args: [currentUserAddress as Address, expenseAccountEip712Address.value]
-  })
+//   const allowance = await readContract(config, {
+//     address: tokenAddress as Address,
+//     abi: ERC20ABI,
+//     functionName: 'allowance',
+//     args: [currentUserAddress as Address, expenseAccountEip712Address.value]
+//   })
 
-  const currentAllowance = allowance ? allowance.toString() : 0n
-  if (Number(currentAllowance) < Number(_amount)) {
-    approve({
-      address: tokenAddress as Address,
-      abi: ERC20ABI,
-      functionName: 'approve',
-      args: [expenseAccountEip712Address.value, _amount]
-    })
-  } else {
-    try {
-      const args = [
-        tokenRecipient.value,
-        _amount,
-        {
-          ...budgetLimit,
-          budgetData: budgetLimit.budgetData.map((item: BudgetData) => ({
-            ...item,
-            value: item.budgetType === 0 ? item.value : BigInt(Number(item.value) * 1e6)
-          }))
-        },
-        signatureToTransfer.value
-      ]
-      const data = encodeFunctionData({
-        abi: expenseAccountABI,
-        functionName: 'transfer',
-        args
-      })
-      await estimateGas(config, {
-        to: expenseAccountEip712Address.value,
-        data
-      })
-      executeExpenseAccountTransfer({
-        address: expenseAccountEip712Address.value,
-        abi: expenseAccountABI,
-        functionName: 'transfer',
-        args
-      })
-    } catch (error) {
-      log.error('Error in transferErc20Token:', error)
-      addErrorToast(parseError(error, expenseAccountABI as Abi))
-      transferERC20loading.value = false
-      isLoadingTransfer.value = false
-    }
-  }
-}
-//#endregion
+//   const currentAllowance = allowance ? allowance.toString() : 0n
+//   if (Number(currentAllowance) < Number(_amount)) {
+//     approve({
+//       address: tokenAddress as Address,
+//       abi: ERC20ABI,
+//       functionName: 'approve',
+//       args: [expenseAccountEip712Address.value, _amount]
+//     })
+//   } else {
+//     try {
+//       const args = [
+//         tokenRecipient.value,
+//         _amount,
+//         {
+//           ...budgetLimit,
+//           budgetData: budgetLimit.budgetData.map((item: BudgetData) => ({
+//             ...item,
+//             value: item.budgetType === 0 ? item.value : BigInt(Number(item.value) * 1e6)
+//           }))
+//         },
+//         signatureToTransfer.value
+//       ]
+//       const data = encodeFunctionData({
+//         abi: expenseAccountABI,
+//         functionName: 'transfer',
+//         args
+//       })
+//       await estimateGas(config, {
+//         to: expenseAccountEip712Address.value,
+//         data
+//       })
+//       executeExpenseAccountTransfer({
+//         address: expenseAccountEip712Address.value,
+//         abi: expenseAccountABI,
+//         functionName: 'transfer',
+//         args
+//       })
+//     } catch (error) {
+//       log.error('Error in transferErc20Token:', error)
+//       addErrorToast(parseError(error, expenseAccountABI as Abi))
+//       transferERC20loading.value = false
+//       isLoadingTransfer.value = false
+//     }
+//   }
+// }
+// //#endregion
 
-//#region Watchers
-watch(newExpenseData, (newData) => {
-  if (newData) {
-    // console.log(`newexpenseData: `, newData)
-  }
-})
-watch(isConfirmingTransfer, async (isConfirming, wasConfirming) => {
-  if (!isConfirming && wasConfirming && isConfirmedTransfer.value) {
-    addSuccessToast('Transfer Successful')
-    // await refetchBalances()
-    transferModal.value = false
-    transferERC20loading.value = false
-    await expenseDataStore.fetchAllExpenseData()
-  }
-})
-watch(errorTransfer, (newVal) => {
-  if (errorTransfer.value) {
-    transferERC20loading.value = false
-    isLoadingTransfer.value = false
-    log.error(parseError(newVal))
-    addErrorToast('Failed to transfer')
-  }
-})
-watch(isConfirmingApprove, (newIsConfirming, oldIsConfirming) => {
-  if (!newIsConfirming && oldIsConfirming && isConfirmedApprove.value) {
-    addSuccessToast('Approval granted successfully')
-    transferERC20loading.value = false
-    transferErc20Token()
-  }
-})
-watch(approveError, () => {
-  if (approveError.value) {
-    transferERC20loading.value = false
-    log.error(parseError(approveError.value))
-    addErrorToast('Failed to approve token spending')
-  }
-})
+// //#region Watchers
+// watch(isConfirmingTransfer, async (isConfirming, wasConfirming) => {
+//   if (!isConfirming && wasConfirming && isConfirmedTransfer.value) {
+//     addSuccessToast('Transfer Successful')
+//     transferModal.value = false
+//     transferERC20loading.value = false
+//     await expenseDataStore.fetchAllExpenseData()
+//   }
+// })
+// watch(errorTransfer, (newVal) => {
+//   if (errorTransfer.value) {
+//     transferERC20loading.value = false
+//     isLoadingTransfer.value = false
+//     log.error(parseError(newVal))
+//     addErrorToast('Failed to transfer')
+//   }
+// })
+// watch(isConfirmingApprove, (newIsConfirming, oldIsConfirming) => {
+//   if (!newIsConfirming && oldIsConfirming && isConfirmedApprove.value) {
+//     addSuccessToast('Approval granted successfully')
+//     transferERC20loading.value = false
+//     transferErc20Token()
+//   }
+// })
+// watch(approveError, () => {
+//   if (approveError.value) {
+//     transferERC20loading.value = false
+//     log.error(parseError(approveError.value))
+//     addErrorToast('Failed to approve token spending')
+//   }
+// })
 //#endregion
 </script>
